回滚硬套 asm + 第 10~13 轮回归成果（md 1426）
- 用户新规矩（2026-08-11）：禁止硬套内联 asm，仅允许纯 nop 对齐；
  到头函数记录 caliber_issues.csv（REMAIN_asm_rollback），不再用 asm 强追
- 回滚 9 服务 13 处 asm 强制（shutdown 死比较族、ChannelService memset/ctor、
  ScriptThread 死比较、ArchiveLog memcpy、relay UDPHandler/Script asm-goto、
  CashObject test/sete），还原为语义版本
- 手册 identical_brief_for_agents.md 新增禁止硬套 asm 条款
- 第 10~13 轮类粒度子分解并行回归：md 1568→1426（回滚前 1428）
- 权威口径：先 gen_report_manifest 再 gen_function_md 全量重生成

+++ dnf_decompile/source/DNFServer/GameServer/Monitor/DNFProhibitUser.cpp
@@ -59,7 +59,7 @@ bool CDNFProhibitUser::IsTimeOutConnectable()
 {
     ((RA_U16<4>*)this)->v =
         (unsigned short)(((RA_U16<4>*)this)->v - 1);
-    if (((RA_U16<4>*)this)->v <= 0)
+    if ((short)(((RA_U16<4>*)this)->v) <= 0)
     {
         return 1;
     }
